+++ sdk/verifiedTransfer/index.js
@@ -1,0 +1,4 @@
+const verifiedTransfer = require('./verifiedTransfer.js');
+
+exports.create = verifiedTransfer.create;
+exports.VerifiedTransfer = verifiedTransfer.VerifiedTransfer;