+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- VerifiedAccount resource
+- VerifiedTransfer resource
 
 ## [2.39.1] - 2026-03-25
 ### Fixed